refactor image view module structure
