--- inappropriate_image_viewer_library/src/main/java/com/sanaa/inappropriate_image_viewer_library/domain/Classification.kt
@@ -1,6 +1,0 @@
-package com.sanaa.inappropriate_image_viewer_library.domain
-
-internal data class Classification(
-    val label: String,
-    val score: Float,
-)
